1. pifUart 기능 개선에 따른 예제 수정함.
2. pifProtocol 예제는 삭제함.

+++ arduino_nano/exDotMatrixS_SR/appMain.cpp
@@ -18,7 +18,7 @@ const uint8_t font8x8_basic[10][8] = {
 };
 
 
-static uint16_t _taskDotMatrixTest(PifTask* p_task)
+static uint32_t _taskDotMatrixTest(PifTask* p_task)
 {
 	static int blink = 0;
 	static int index = 0;
